s1250124 added 2 files on class3_10.

+++ AdjacencyList.java
@@ -4,24 +4,24 @@ import java.util.HashMap;
 import java.util.Set;
 class AdjacencyList{
     /*adlist has Map(Index,distance)*/
-    private Map<Integer,Integer> adlist;
+    private Map<Integer,Double> adlist;
     AdjacencyList(){
-	adlist = new HashMap<Integer,Integer>();
+	adlist = new HashMap<Integer,Double>();
     }
-    public void put(int key,int d){
-	adlist.put((Integer)key,(Integer)d);
+    public void put(int key,double d){
+	adlist.put((Integer)key,(Double)d);
     }
-    public int get(int key){
-	int i=adlist.get((Integer)key);
-	if((Integer)i!=null){
-	    return (int)i;
+    public double get(int key){
+	double i=adlist.get((Integer)key);
+	if((Double)i!=null){
+	    return (double)i;
 	}else{
-	    return -1;
+	    return -1.0;
 	}
     }
     public boolean remove(int key){
-	int i=adlist.remove((Integer)key);
-	if((Integer)i!=null){
+	double i=adlist.remove((Integer)key);
+	if((Double)i!=null){
 	    return true;
 	}else{
 	    return false;
@@ -31,12 +31,12 @@ class AdjacencyList{
 	return adlist.containsKey((Integer)key);
     }
     public HashMap deepcopy(){
-	return new HashMap<Integer,Integer>(adlist);
+	return new HashMap<Integer,Double>(adlist);
     }
     public Set<Integer> keySet(){
 	return adlist.keySet();
     }
-    public Set<Entry<Integer,Integer>> entrySet(){
+    public Set<Entry<Integer,Double>> entrySet(){
 	return adlist.entrySet();
     }//usage(blog): https://qiita.com/kei2100/items/0ce97733c92fdcb9c5a9
